Moved IGrainControlExtension
