+++ src/Outkeep.Grains.Interfaces/Governance/IGrainControlExtension.cs
@@ -1,7 +1,7 @@
 ﻿using Orleans.Runtime;
 using System.Threading.Tasks;
 
-namespace Outkeep.Grains
+namespace Outkeep.Grains.Governance
 {
     public interface IGrainControlExtension : IGrainExtension
     {
