+++ src/Outkeep.Grains/GrainControlExtension.cs
@@ -1,4 +1,5 @@
 ﻿using Orleans.Runtime;
+using Outkeep.Grains.Governance;
 using System;
 using System.Threading.Tasks;
